changed segtree_iterative for segtree, for non-commutative support

+++ data_structures/segment_tree.cpp
@@ -1,13 +1,12 @@
-// works in all cases, even non-commutative operations (see test2)
 typedef ll node;
 #define oper min
 #define NEUT INF
-struct STree{ // segment tree for min over long long integers
+struct STree{ // segment tree for min over ll integers
 	int n; vector<node>t;
 	STree(int n):n(n),t(2*n+5,NEUT){}
 	void init(vector<node> &a){
 		fore(i,0,n)t[n+i]=a[i];
-		for(ll i=n-1;i>0;i--)t[i]=oper(t[2*i],t[2*i+1]);
+		for(int i=n-1;i>0;i--)t[i]=oper(t[2*i],t[2*i+1]);
 	}
 	void upd(int p, node v){
 		for(p+=n,t[p]=v;p>1;p>>=1)p|=1,t[p>>1]=oper(t[p^1],t[p]);
